updating ActivityServiceImpl method deleteActivity() to throw exception with test

+++ service/src/main/java/com/demo/rest/api/service/ActivityServiceImpl.java
@@ -45,6 +45,9 @@ public class ActivityServiceImpl implements ActivityService {
     @Override
     @Transactional
     public void deleteActivity(Long id) {
+        Optional<Activity> activity = activityDao.findById(id);
+        if (activity.isEmpty())
+            throw new ActivityNotFoundException("Activity with id=" + id + " not found!");
         activityDao.deleteById(id);
     }
 }
